move completely independent tests
